quickstarts/fastapi-mongo/requirements.txt
fastapi
uvicorn
motor
web3
python-dotenv
starlette
